requirements.txt
streamlit
tensorflow
torch
torchvision
pillow
numpy
pandas
matplotlib
plotly
passlib
python-dotenv
requests
google-generativeai
timm
